requirements.txt
flask>=3.0.0
openpyxl>=3.1.0
python-pptx>=0.6.23
